requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
fastapi==0.112.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
outcome==1.3.0.post0
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
starlette==0.37.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
websocket-client==1.8.0
wsproto==1.2.0
